Extract syscall_close() handler
